multi: move ChannelUpdate validate methods to netann
